+++ AutomatedLab/AutomatedLab.psm1
@@ -548,6 +548,7 @@ function Import-Lab
         try
         {
             $Script:data.Disks = $importMethodInfo.Invoke($null, $Script:data.DiskDefinitionFiles[0].Path)
+            $Script:data.Disks = Get-LabVHDX -All
 
             if ($Script:data.DiskDefinitionFiles.Count -gt 1)
             {
@@ -1486,7 +1487,7 @@ function Remove-Lab
                         }
                         else
                         {
-                            Write-ScreenInfo "Disk '$($disk.Path)' does not exist"
+                            Write-ScreenInfo "Disk '$($disk.Path)' does not exist" -Type Verbose
                         }
                     }
                 }