+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - When Az credentials were expired, lab resources were not removed
 - When calculating the required free space on disk for labs on Hyper-V, also VMs
   that are enabled for SkipDeployment were taken into account. This is now fixed.
+- The network interface name was not assigned if there is only one interface defined for a VM.
 
 ## 5.38.0 (2021-07-14)