+++ CHANGELOG.md
@@ -1,10 +1,30 @@
 # Changelog
 
-## Unreleased
+## Unreleased (yyyy-MM-dd)
 
 ### Enhancements
 
 ### Bugs
+
+## 5.39.0 (2021-08-20)
+
+### Enhancements
+
+- Exchange links updated
+- NuGet custom role distributes nuget.exe on lab VMs and registers repository
+- No more validating non-deployed roles
+- AutomatedLab.Recipe can now also use an Azure Storage account to store snippets and samples. No custom roles at the moment though.
+- Support for multiple storage nodes, controlled by new role property
+  'StorageTarget' for role 'FailoverNode'
+- The VHDx files for additional disks are now stored in the VM's path and no longer in the
+  disks folder.
+
+### Bugs
+
+- When Az credentials were expired, lab resources were not removed
+- When calculating the required free space on disk for labs on Hyper-V, also VMs
+  that are enabled for SkipDeployment were taken into account. This is now fixed.
+- The network interface name was not assigned if there is only one interface defined for a VM.
 
 ## 5.38.0 (2021-07-14)